+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - New tech cards
   - Titan Armour Tech
+  - Titan Ammunition Tech
 
 ## v5.66.2 - 2024-07-28